refactor: frame과 관련된 클래스를 하나의 패키지로 묶음
