+++ src/main/java/bowling/domain/Game.java
@@ -1,5 +1,8 @@
 package bowling.domain;
 
+import bowling.domain.frame.Frame;
+import bowling.domain.frame.PinCount;
+
 import java.util.List;
 
 public class Game {
