+++ src/main/java/bowling/domain/frame/NormalFrame.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.frame;
 
 import java.util.List;
 import java.util.Optional;
