+++ src/test/java/bowling/domain/GameTest.java
@@ -1,5 +1,6 @@
 package bowling.domain;
 
+import bowling.domain.frame.PinCountTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
