+++ src/test/java/bowling/domain/frame/PinCountTest.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.frame;
 
 import bowling.domain.exception.OutOfRangeArgumentException;
 import org.junit.jupiter.api.DisplayName;
